Added stats and cleaned up UI

+++ DrPhil Controller/DrPhil Controller/System/ObjectStructures/DailyStats.swift
@@ -19,6 +19,6 @@ struct DailyStats {
 }
 
 struct ResourceMetrics {
-    var battery : Double
-    var disinfectant : Double
+    var battery : Float
+    var disinfectant : Float
 }
